fix: move LogHighlighter from Gui to Core
move loghighlighter to allow use in Gui and Core classes
